refactor: styles and configs
