+++ src/app/authentication/login/index.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 import { LoginForm } from "@/features/authentication/components/LoginForm";
 import {
   LoginWrapper,
