+++ src/app/dashboard/home/index.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Row, Col } from "antd";
 
 import { DashboardHeader } from "@/features/dashboard/components/DashboardHeader";
