+++ src/features/authentication/components/LoginForm.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Form, Input, Button } from "antd";
 import { UserOutlined, LockOutlined } from "@ant-design/icons";
