+++ src/features/dashboard/components/StatCard.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import { Typography } from "antd";
 
 import {
