+++ src/main.tsx
@@ -1,5 +1,6 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
+
 import App from "@/App.tsx";
 
 createRoot(document.getElementById("root")!).render(
