+++ src/routes/private-routes.ts
@@ -1,4 +1,5 @@
 import { lazy } from "react";
+
 import { PRIVATE_ROUTES } from "./paths";
 
 const { DASHBOARD } = PRIVATE_ROUTES;
